## Popup wpm level <li> logic <li> styling
The logic for changing from one animal to another,
based on the result wpm was added.
Styling was also added to make sure when the window is scaled down
the card is adjusted.

## Result layout
<li> styling
<li> refactored

This was entirely changed
due to the previous method not being at all responsive
and clean.

This new implementation uses Material Ui **(Grid - Paper)**
Three breakpoints are set [sm, md, xs] for responsivness

<hr>

All in all, we got out some of the most annoying things ever

See issue: #6.